\name{lmf-package}
\alias{lmf-package}
\docType{package}
\title{
Functions for estimation and inference of selection in age-structured populations
}
\description{
This R package estimates directional and fluctuating selection in
age-structured populations and provides methods for statisical inference using
the procedures developed by Steinar Engen et al. 2012. Also compatible with
data without age-structure.
}
\details{
\tabular{ll}{
Package: \tab lmf\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-04-10\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
}
\author{
Thomas Kvalnes, Centre for Conservation Biology, Norwegian University of
Science and Technology, N-7491 Trondheim, Norway

Maintainer: Thomas Kvalnes <thomas.kvalnes@bio.ntnu.no>
}
\references{
Engen, S., Saether, B.-E., Kvalnes, T. and Jensen, H. 2012. Estimating fluctuating
selection in age-structured populations. Journal of Evolutionary Biology, 00, 000-000.
}
